+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/backpacks/Cooler.java
@@ -1,12 +1,13 @@
-package io.github.thebusybiscuit.slimefun4.implementation.items.food;
+package io.github.thebusybiscuit.slimefun4.implementation.items.backpacks;
 
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
-import io.github.thebusybiscuit.slimefun4.implementation.items.backpacks.SlimefunBackpack;
+import io.github.thebusybiscuit.slimefun4.implementation.items.food.Juice;
 import io.github.thebusybiscuit.slimefun4.implementation.listeners.CoolerListener;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
 /**
@@ -24,6 +25,12 @@ public class Cooler extends SlimefunBackpack {
 
     public Cooler(int size, Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(size, category, item, recipeType, recipe);
+    }
+
+    @Override
+    public boolean isItemAllowed(ItemStack item, SlimefunItem itemAsSlimefunItem) {
+        // A Cooler only allows Juices
+        return itemAsSlimefunItem instanceof Juice;
     }
 
 }
