+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/backpacks/SlimefunBackpack.java
@@ -1,5 +1,6 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.backpacks;
 
+import org.bukkit.Material;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 
@@ -40,6 +41,27 @@ public class SlimefunBackpack extends SimpleSlimefunItem<ItemUseHandler> {
      */
     public int getSize() {
         return size;
+    }
+
+    /**
+     * This method returns whether a given {@link ItemStack} is allowed to be stored
+     * in this {@link SlimefunBackpack}.
+     * 
+     * @param item
+     *            The {@link ItemStack} to check for
+     * 
+     * @param itemAsSlimefunItem
+     *            The same {@link ItemStack} as a {@link SlimefunItem}, might be null
+     * 
+     * @return Whether the given {@link ItemStack} is allowed to be put into this {@link SlimefunBackpack}
+     */
+    public boolean isItemAllowed(ItemStack item, SlimefunItem itemAsSlimefunItem) {
+        // Shulker Boxes are not allowed!
+        if (item.getType() == Material.SHULKER_BOX || item.getType().toString().endsWith("_SHULKER_BOX")) {
+            return false;
+        }
+
+        return !(itemAsSlimefunItem instanceof SlimefunBackpack);
     }
 
     @Override
